# d2s.c: Further improve x86 perf.